refactor(ui): move syntax highlighters into subdirectory
I will be adding some search / replace logic.
I will likely add another top-level highlighter that composes the patterned highlighter and the search one.
