+++ ui/highlight/syntax/patternedhiglighter.cpp
