+++ ui/highlight/syntax/rules_clike.h
